Start to refactor test suite; update for https
Move the test suite files around so that they can be run.

Start to update linked.data.gov.au.json to use https-based
URLs.
