+++ test-suite/count-patterns.sh
