+++ test-suite/linked.data.gov.au-vocabularies.py
