+++ test-suite/tests/catalogue_tester.py
