+++ test-suite/tests/rules_template/Rule.py
